Decouple storm config into StormExecutionRuntime from StormEnvironment
